requirements.txt
# Python library dependencies for workday_data_extractor
requests==2.31.0
python-dotenv==1.0.0